+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Deployment modal - Submit button show loading when clicked
   - Menubar buttons - Capitalize all of them
   - Modal headers - Capitalize all of them
   - Format operation - some action names are lower case